+++ assets/src/components/account/groups/GroupCreate.tsx
@@ -5,13 +5,13 @@ import { Button, Modal, ValidatedInput } from '@pluralsh/design-system'
 
 import isEmpty from 'lodash/isEmpty'
 
-import { appendConnection, updateCache } from '../../utils/graphql'
+import { appendConnection, updateCache } from '../../../utils/graphql'
 
-import { GqlError } from '../utils/Alert'
+import { GqlError } from '../../utils/Alert'
 
-import { CREATE_GROUP, GROUPS_Q } from './queries'
+import { CREATE_GROUP, GROUPS_Q } from '../queries'
 
-export function CreateGroup({ q }: any) {
+export default function GroupCreate({ q }: {q: string}) {
   const [open, setOpen] = useState(false)
   const [name, setName] = useState('')
   const [description, setDescription] = useState('')
