+++ assets/src/components/account/groups/GroupView.tsx
@@ -1,0 +1,15 @@
+import { Box } from 'grommet'
+
+import GroupMembers from './GroupMembers'
+
+export default function GroupView({ group }: any) {
+  return (
+    <Box
+      fill
+      pad={{ bottom: 'small' }}
+      gap="small"
+    >
+      <GroupMembers group={group} />
+    </Box>
+  )
+}
